crc-64-avro code moded to independent file
